restructure Account and add NotFunded message
